app/requirements.txt
fastapi
uvicorn[standard]
requests
minio
weaviate-client
unstructured
httpx
pytest
pytest-asyncio
python-dotenv
